requirements.txt
asgiref==3.5.0
Django==4.0.2
django-environ==0.8.1
django-extensions==3.1.5
django-filter~=21.1
django-rest-framework-roles==0.6
djangorestframework==3.13.1
drf-extensions==0.7.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
